change representiative to member
I'm not a member of an underrepresented group, so please take this with a grain of salt. My impression is that being a representative is a lot more work than being a member of a group. 

I certainly wouldn't want to be seen as the representative of middle aged white guys.

This is offered as an improvement.  Thanks for all your other work on curl!